refactor(RecordForm): rename inputValidators to input
